# perf: Document embedding getter