This PR doubles the performance for retrieving embeddings using `doc.embedding` without any change on the API.

This is achieved removing the instantiation of  a `NdArray`  created only to get it's `.value` (which is a numpy array), essentially doubling performance.

Benchmark

```
d = Document(embedding = np.random.rand((128)))
d.embedding
```

- This PR
      ```
      %timeit aux = d.embedding
      2.4 µs ± 50.6 ns per loop
      ```

- Master
      ```
       %timeit aux = d.embedding
       6.2 µs ± 173 ns per loop 
      ```